# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FoodCreateRequest": {
        "properties": {
          "brand": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "caffeine_per_100g": {
            "default": 0,
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Caffeine Per 100G",
            "type": "number"
          },
          "carbs_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Carbs Per 100G",
            "type": "number"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "fat_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Fat Per 100G",
            "type": "number"
          },
          "kcal_per_100g": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Kcal Per 100G",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "protein_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Protein Per 100G",
            "type": "number"
          },
          "serving_unit": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Unit"
          },
          "serving_weight_g": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Weight G"
          },
          "sodium_per_100g": {
            "default": 0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Sodium Per 100G",
            "type": "number"
          },
          "sugar_per_100g": {
            "default": 0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sugar Per 100G",
            "type": "number"
          }
        },
        "required": [
          "name",
          "kcal_per_100g",
          "protein_per_100g",
          "fat_per_100g",
          "carbs_per_100g"
        ],
        "title": "FoodCreateRequest",
        "type": "object"
      },
...
      "FoodMatchResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FoodMatchCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "preselected_food_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preselected Food Item Id"
          },
          "status": {
            "enum": [
              "no_match",
              "ambiguous",
              "review"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "preselected_food_item_id",
          "candidates"
        ],
        "title": "FoodMatchResponse",
        "type": "object"
      },
      "FoodResponse": {
        "properties": {
          "brand": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "caffeine_per_100g": {
            "default": 0,
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Caffeine Per 100G",
            "type": "number"
          },
          "carbs_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Carbs Per 100G",
            "type": "number"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_revision": {
            "title": "Current Revision",
            "type": "string"
          },
          "density_g_per_ml": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Density G Per Ml"
          },
          "fat_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Fat Per 100G",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kcal_per_100g": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Kcal Per 100G",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nutrition_basis_quantity": {
            "title": "Nutrition Basis Quantity",
            "type": "number"
          },
          "nutrition_basis_unit": {
            "enum": [
              "g",
              "ml",
              "serving"
            ],
            "title": "Nutrition Basis Unit",
            "type": "string"
          },
          "nutrition_complete": {
            "title": "Nutrition Complete",
            "type": "boolean"
          },
          "protein_per_100g": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Protein Per 100G",
            "type": "number"
          },
          "serving_unit": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Unit"
          },
          "serving_weight_g": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Weight G"
          },
          "sodium_per_100g": {
            "default": 0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Sodium Per 100G",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "sugar_per_100g": {
            "default": 0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sugar Per 100G",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kcal_per_100g",
          "protein_per_100g",
          "fat_per_100g",
          "carbs_per_100g",
          "id",
          "source",
          "source_reference",
          "current_revision",
          "nutrition_basis_unit",
          "nutrition_basis_quantity",
          "density_g_per_ml",
          "nutrition_complete",
          "created_at",
          "updated_at"
        ],
        "title": "FoodResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogResponse": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "caffeine": {
            "title": "Caffeine",
            "type": "number"
          },
          "carbs": {
            "title": "Carbs",
            "type": "number"
          },
          "catalog_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Revision"
          },
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "custom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "fat": {
            "title": "Fat",
            "type": "number"
          },
          "food_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Item Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kcal": {
            "title": "Kcal",
            "type": "number"
          },
          "log_date": {
            "format": "date",
            "title": "Log Date",
            "type": "string"
          },
          "meal_type": {
            "$ref": "#/components/schemas/MealType"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "nutrition_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nutrition Source"
          },
          "nutrition_source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nutrition Source Reference"
          },
          "protein": {
            "title": "Protein",
            "type": "number"
          },
          "sodium": {
            "title": "Sodium",
            "type": "number"
          },
          "sugar": {
            "title": "Sugar",
            "type": "number"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "client_id",
          "log_date",
          "meal_type",
          "food_item_id",
          "custom_name",
          "amount",
          "unit",
          "kcal",
          "protein",
          "fat",
          "carbs",
          "sugar",
          "sodium",
          "caffeine",
          "note",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "LogResponse",
        "type": "object"
      },
...
      "MealType": {
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack",
          "drink"
        ],
        "title": "MealType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NutriPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/foods": {
      "get": {
        "operationId": "search_foods_api_v1_foods_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FoodResponse"
                  },
                  "title": "Response Search Foods Api V1 Foods Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Foods",
        "tags": [
          "diet"
        ]
      },
      "post": {
        "operationId": "add_food_api_v1_foods_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Food",
        "tags": [
          "diet"
        ]
      }
    },
    "/api/v1/foods/match": {
      "get": {
        "operationId": "match_foods_api_v1_foods_match_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "brand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Match Foods",
        "tags": [
          "diet"
        ]
      }
    },
...
    "/api/v1/logs": {
      "get": {
        "operationId": "read_logs_api_v1_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LogResponse"
                  },
                  "title": "Response Read Logs Api V1 Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Logs",
        "tags": [
          "diet"
        ]
      },
      "post": {
        "operationId": "add_log_api_v1_logs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Idempotent replay"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Key conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Log",
        "tags": [
          "diet"
        ]
      }
    },
    "/api/v1/logs/{log_id}": {
      "delete": {
        "operationId": "remove_log_api_v1_logs__log_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Log Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Log",
        "tags": [
          "diet"
        ]
      },
      "get": {
        "operationId": "read_log_api_v1_logs__log_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Log Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Log",
        "tags": [
          "diet"
        ]
      },
      "put": {
        "operationId": "update_log_api_v1_logs__log_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Log Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Log",
        "tags": [
          "diet"
        ]
      }
    },
...
    "/api/v1/stats/daily": {
      "get": {
        "operationId": "read_daily_summary_api_v1_stats_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "summary_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Summary Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Daily Summary",
        "tags": [
          "diet"
        ]
      }
    },
    "/api/v1/sync/changes": {
      "get": {
        "operationId": "read_sync_changes_api_v1_sync_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Sync Changes",
        "tags": [
          "diet"
        ]
      }
    },
...
  }
}